+++ sdk/appcenter-identity/src/androidTest/java/com/microsoft/appcenter/identity/storage/TokenStorageTest.java
@@ -1,4 +1,4 @@
-package com.microsoft.appcenter.utils.context;
+package com.microsoft.appcenter.identity.storage;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
