+++ sdk/appcenter-identity/src/main/java/com/microsoft/appcenter/identity/storage/TokenStorageFactory.java
@@ -1,4 +1,4 @@
-package com.microsoft.appcenter.utils.context;
+package com.microsoft.appcenter.identity.storage;
 
 import android.content.Context;
